move graph.h into graph
