+++ sharkfuser/include/fusilli/asm_emitter.h
@@ -30,7 +30,7 @@
 
 #include "fusilli/attributes/tensor_attributes.h"
 #include "fusilli/extras.h"
-#include "fusilli/graph.h"
+#include "fusilli/graph/graph.h"
 #include "fusilli/node/conv_node.h"
 #include "fusilli/types.h"
