+++ sharkfuser/include/fusilli/backend.h
@@ -19,6 +19,7 @@
 #include <vector>
 
 namespace fusilli {
+
 // Where do we want the generated code to run?
 enum class Backend {
   CPU,
@@ -45,4 +46,5 @@ static const std::unordered_map<Backend, std::vector<std::string>>
 };
 
 } // namespace fusilli
+
 #endif // FUSILLI_BACKEND_H
